Refactor the caroussel
