+++ src/components/assets/styles/catcard.scss
@@ -1,18 +1,6 @@
 .wrapper {
   display: flex;
-  gap: 1em;
-}
-
-.category-container {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-template-rows: repeat(5, 1fr);
-  grid-column-gap: 0;
-  grid-row-gap: 0;
-  background-repeat: no-repeat;
-  text-align: center;
-  margin: 5rem auto 1rem;
-  width: 100%;
+  overflow: hidden;
 }
 
 a {
@@ -24,12 +12,13 @@ a {
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  width: 80%;
+  width: 100%;
   box-shadow: 0 2px 2px rgba(0, 0, 0, 0.15);
   color: #5d5e5e;
   border-radius: 4px;
   border: 1px solid #e7e7e7;
   background-color: #fff;
+  overflow: hidden;
   padding: 1rem;
   text-align: center;
   margin: auto;
@@ -65,14 +54,7 @@ h3 {
 /* Desktop Version */
 @media screen and (min-width: 768px) {
   .card {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    box-shadow: 0 2px 2px rgba(245, 145, 145, 0.15);
-    color: #5d5e5e;
-    border-radius: 4px;
     border: 2px solid #e7e7e7;
-    background-color: #fff;
     width: 280px;
     height: 280px;
     margin: 0;
