+++ src/components/assets/styles/catpage.scss
@@ -1,0 +1,6 @@
+.category-container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100vh;
+}
