--- src/components/categories/catpage.scss
@@ -1,24 +1,0 @@
-.category-container {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: wrap;
-  background-repeat: no-repeat;
-  text-align: center;
-  margin: 5rem auto 1rem;
-  width: 100%;
-  height: 100%;
-}
-
-@media screen and (min-width: 768px) {
-  .category-container {
-    display: flex;
-    flex-wrap: wrap;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    padding: 2rem;
-    margin: 5rem auto;
-    width: 90%;
-    height: 70vh;
-  }
-}
